+++ lib/ui/providers/app_providers.dart
@@ -1,0 +1,29 @@
+import 'package:places/ui/providers/add_photo_provider.dart';
+import 'package:places/ui/providers/add_sight_provider.dart';
+import 'package:places/ui/providers/filter_provider.dart';
+import 'package:places/ui/providers/search_provider.dart';
+import 'package:places/ui/providers/sight_list_provider.dart';
+import 'package:places/ui/providers/visiting_screen_provider.dart';
+import 'package:provider/provider.dart';
+import 'package:provider/single_child_widget.dart';
+
+List<SingleChildWidget> appProviders = [
+  ChangeNotifierProvider<AddSightProvider>(
+    create: (_) => AddSightProvider(),
+  ),
+  ChangeNotifierProvider<AddPhotoProvider>(
+    create: (_) => AddPhotoProvider(),
+  ),
+  ChangeNotifierProvider<FilterProvider>(
+    create: (_) => FilterProvider(),
+  ),
+  ChangeNotifierProvider<SearchProvider>(
+    create: (_) => SearchProvider(),
+  ),
+  ChangeNotifierProvider<SightListProvider>(
+    create: (_) => SightListProvider(),
+  ),
+  ChangeNotifierProvider<VisitingProvider>(
+    create: (_) => VisitingProvider(),
+  ),
+];
